medichatbot/requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
cassio
tiktoken
gradio

-e .